refactor(workspace): make restore-cohort region authoritative, drop appStore reducers
